fix handling of request failures (fixes #45)
Previously, in src/tutor.coffee `get` would invoke the success callback
*before* checking `err`. The following code would therefore, in case of
an error, attempt to read the "legality" property of undefined:

    gatherer.printings details, get (data) -> {legality} = data

This line, which appeared several times, was also erroneous:

    err ?= new Error 'unexpected status code' unless res.statusCode is 200

It was intended to produce the following JavaScript:

    if (err == null) {
      if (res.statusCode !== 200) {
        err = new Error('unexpected status code');
      }
    }

The checks were actually performed the other way around, so this, too,
could attempt to read a property of undefined.

This commit centralizes error handling in addition to fixing these bugs.